+++ modules/Accounting/StaffBalances.php
@@ -1,19 +1,18 @@
 <?php
 DrawHeader( ProgramTitle() );
 
-$extra['SELECT'] = ",(COALESCE(
-	(SELECT SUM(f.AMOUNT)
-		FROM accounting_salaries f
-		WHERE f.STAFF_ID=s.STAFF_ID
-		AND f.SCHOOL_ID='" . UserSchool() . "'
-		AND f.SYEAR=s.SYEAR),0)
-	-COALESCE(
-	(SELECT SUM(p.AMOUNT)
-		FROM accounting_payments p
-		WHERE p.STAFF_ID=s.STAFF_ID
-		AND p.SCHOOL_ID='" . UserSchool() . "'
-		AND p.SYEAR=s.SYEAR)
-	,0)) AS BALANCE";
+// @since 12.9 SQL performance: replace subqueries with LEFT JOINs
+// SQL remove SCHOOL_ID join: fix when "Search All Schools" is checked
+$extra['SELECT'] = ',COALESCE(f.TOTAL,0)-COALESCE(p.TOTAL,0) AS BALANCE';
+
+$extra['FROM'] = ' LEFT JOIN (SELECT STAFF_ID,SYEAR,SUM(AMOUNT) AS TOTAL
+	FROM accounting_salaries
+	GROUP BY STAFF_ID,SYEAR) f ON f.STAFF_ID=s.STAFF_ID
+	AND f.SYEAR=s.SYEAR
+LEFT JOIN (SELECT STAFF_ID,SYEAR,SUM(AMOUNT) AS TOTAL
+	FROM accounting_payments
+	GROUP BY STAFF_ID,SYEAR) p ON p.STAFF_ID=s.STAFF_ID
+	AND p.SYEAR=s.SYEAR';
 
 $extra['columns_after'] = [ 'BALANCE' => _( 'Balance' ) ];