+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 12.9
 - Fix PHP warning undefined index when school was deleted in School.php & Schools.php
 - Security fix Request Forgery: remove modfunc= in links in Modules.php & MarkDownHTML.fnc.php
 - Performance: do not sort `functions/*.php` files in Warehouse.php
+- SQL performance: replace subqueries with LEFT JOINs in StaffWidget/StaffBalance.php, Widget/Absences.php, Widget/Balance.php, Widget/CpAbsences.php, StaffBalances.php, StudentSummary.php, StudentBalances.php & Export.php
 
 Changes in 12.8
 ---------------